### 2023-03-10 #5241 @editorialbot
[PRE REVIEW]: Zoobot: Adaptable Deep Learning Models for Galaxy Morphology
**Submitting author:** @mwalmsley (<a href="http://orcid.org/0000-0002-6408-4181">Mike Walmsley</a>)
**Repository:** https://github.com/mwalmsley/zoobot
**Branch with paper.md** (empty if default branch): docs
**Version:** v1.0.0
