+++ paper/paper.bib
@@ -42,6 +42,7 @@ especially in the low-label transfer regime (+6% accuracy with 750 labels).},
    volume = {509},
    url = {https://arxiv.org/abs/2102.08414},
    year = {2022},
+   doi = {10.1093/mnras/stab2093}
 }
 
 @article{2011arXiv1110.3193L,
@@ -111,6 +112,7 @@ archivePrefix = {arXiv},
    title = {The DAWES review 10: The impact of deep learning for the analysis of galaxy surveys},
    url = {http://arxiv.org/abs/2210.01813},
    year = {2022},
+   doi = {10.1017/pasa.2022.55}
 }
 
 @article{LeCun2015,
