### 2023-03-10 48bee11
temporarily set DECaLS as paper-to-cite
