### 2023-03-10 #5241 @dfm
@mwalmsley — I'm going to flag this submission for editorial review so that the editorial board can weigh in on the scope, since we [explicitly don't publish pre-trained ML models in JOSS](https://joss.readthedocs.io/en/latest/submitting.html#what-we-mean-by-research-software). This process normally takes about a week, and I'll keep you posted. I understand that `zoobot` also offers features for updating the models for new tasks, but if you have any further context about how it fits into [the JOSS definition of research software](https://joss.readthedocs.io/en/latest/submitting.html#what-we-mean-by-research-software), please feel free to comment on this thread. Thanks for your patience with this process!
