### 2023-03-15 f68b404
drop extra dims
