### 2023-03-20 0ce66b7
adapt for lightning v2, simplify steps
