+++ zoobot/pytorch/training/finetune.py
@@ -105,6 +105,11 @@ class FinetuneableZoobotAbstract(pl.LightningModule):
 
         self.freeze_batchnorm = freeze_batchnorm
 
+        self.train_loss_metric = tm.MeanMetric()
+        self.val_loss_metric = tm.MeanMetric()
+        self.test_loss_metric = tm.MeanMetric()
+
+
         if self.freeze_batchnorm:
             freeze_batchnorm_layers(self.encoder)  # inplace
 
@@ -157,48 +162,65 @@ class FinetuneableZoobotAbstract(pl.LightningModule):
         x = self.head(x)
         return x
 
-    
     def make_step(self, batch):
+        y, y_pred, loss = self.run_step_through_model(batch)
+        return self.step_to_dict(y, y_pred, loss)
+
+    def run_step_through_model(self, batch):
       # part of training/val/test for all subclasses
         x, y = batch
         y_pred = self.forward(x)
-        loss = self.loss(y_pred, y)
-        # {'loss': loss.mean(), 'predictions': y_pred, 'labels': y}
+        loss = self.loss(y_pred, y)  # must be subclasses and specified
         return y, y_pred, loss
 
+    def step_to_dict(self, y, y_pred, loss):
+        return {'loss': loss.mean(), 'predictions': y_pred, 'labels': y}
 
-    # def on_train_batch_end(self, outputs, *args) -> None:
-    #     self.log("finetuning/train_loss_batch",
-    #              outputs['loss'], on_step=False, on_epoch=True, prog_bar=self.prog_bar)
+    def training_step(self, batch, batch_idx, dataloader_idx=0):
+        return self.make_step(batch)
 
-    def train_epoch_end(self, outputs, *args) -> None:
-        losses = torch.cat([batch_output['loss'].expand(0)
-                           for batch_output in outputs])
-        self.log("finetuning/train_loss",
-                 losses.mean(), prog_bar=self.prog_bar)
+    def validation_step(self, batch, batch_idx, dataloader_idx=0):
+        return self.make_step(batch)
 
-        if hasattr(self, 'train_acc') is not None:
-            self.log("finetuning/train_acc", self.train_acc,
-                     prog_bar=self.prog_bar)  # log here
+    def test_step(self, batch, batch_idx, dataloader_idx=0):
+        return self.make_step(batch)
 
-    def on_validation_epoch_end(self, outputs, *args) -> None:
-        # calc. mean of losses over val batches as val loss
-        losses = torch.FloatTensor([batch_output['loss']
-                                   for batch_output in outputs])
-        self.log("finetuning/val_loss", torch.mean(losses),
-                 prog_bar=self.prog_bar)
+# lighting v2. removed validation_epoch_end(self, outputs)
+# now only has *on_*validation_epoch_end(self)
+# replacing by using explicit torchmetric for loss
+# https://github.com/Lightning-AI/lightning/releases/tag/2.0.0
 
-        if hasattr(self, 'val_acc'):
-            self.log("finetuning/val_acc", self.val_acc,
-                     prog_bar=self.prog_bar)
+# https://torchmetrics.readthedocs.io/en/stable/pages/lightning.html#common-pitfalls
+    def training_step_end(self, step_output):
+        self.train_loss_metric(step_output['loss'])
+        self.log(
+            "finetuning/train_loss", 
+            self.train_loss_metric, 
+            prog_bar=self.prog_bar, 
+            on_step=False,
+            on_epoch=True
+        )
 
-    def on_test_batch_end(self, outputs, *args) -> None:
-        self.log('test/test_loss', outputs['loss'])
-        if hasattr(self, 'test_acc'):
-            self.test_acc(outputs['predictions'], outputs['labels'])
[…]
