### 2023-03-20 c9bf4fb
use on_x_batch_end for logging
