### 2023-03-20 0b1a095
auto strategy by default
