### 2023-03-28 0bbe9e1
add early stopping back, update docs
