### 2023-03-29 65eedf6
color from scratch with fixed channels + vit
