### 2023-03-29 4afe89d
remake core architectures post-alpha-fix
