### 2023-04-24 #5312 @plaplant
@crhea93 thanks for reviewing the software! For issues that should be addressed as part of the review, please open issues on [the upstream repository](https://github.com/mwalmsley/zoobot). Thanks!
