### 2023-04-25 #5312 @crhea93
@plaplant Should the issues have a specific template? I can't find one in the JOSS documentation.
