### 2023-04-25 #5312 @mwalmsley
Hi @crhea93. Thanks for taking the time to review our JOSS submission!

> The requirements file is incorrectly formatted. There is also no reference to the requirements.txt file in the installation.

The `requirements.txt` file is deprecated (the first line is """Deprecated - see setup.py""" and the entire file is commented out). I will remove it now for clarity. Installation dependencies are instead included via `setup.py`, which allows for more configuration (e.g. `zoobot[pytorch]`, `zoobot[tensorflow]`).

> The authors is missing references in the software paper.

Which references are missing? Do you mean missing as in 'should be cited' or 'not linked correctly to bibliography'?

I'm happy to chat directly on here about issues not related to the software itself, or you're welcome to open issues (I don't think a specific structure is required, it's only me reading them ;) )
