### 2023-04-30 #5312 @mwalmsley
Hi reviewers @devanshkv @crhea93. Thank you for your feedback thus far.

I just wanted to highlight a few pieces of code that may be relevant to your remaining checklist items:
- There is a [Colab notebook](https://colab.research.google.com/drive/17bb_KbA2J6yrIm4p4Ue_lEBHMNC1I9Jd?usp=sharing) for finetuning Zoobot. This is aimed at new users but can be used to verify the software works.
- The models can be trained from scratch with using the scripts in /benchmarks to train the models from scratch. This is a manual functional test that new versions of Zoobot still perform well (and hopefully better). Results are uploaded to WandB and I'm happy to provide access to verify, if you like.
- You can use the scripts to train most of the models yourself (though you will need some significant resources and potentially some adaption to your own cluster). All the data is provided in a separate repo, [mwalmsley/galaxy-datasets](github.com/mwalmsley/galaxy-datasets), designed for easy use with ML. This includes GZ2, GZ CANDELS, GZ Hubble, and GZ DECALS. The only training data not yet available via that repo is GZ DESI, which is not original data in this paper - it is being published separately and is currently under review at MNRAS.
