### 2023-05-02 #5312 @plaplant
@crhea93 thanks very much for your review!

@devanshkv are there any outstanding issues that you feel need to be addressed prior to publication? If so, please make an issue on [the upstream repository](https://github.com/mwalmsley/zoobot) and reply to this thread so we're aware. Otherwise, if you feel this is ready for publication, please update your reviewer checklist accordingly. Thanks again for reviewing!
