### 2023-05-02 #5312 @plaplant
@devanshkv I see that there is still an unchecked item on your checklist about verifying the functional claims of the software. Please update your checklist accordingly if you have done this, or first verify the functionality before checking this box. (Per the [JOSS reviewer docs](https://joss.readthedocs.io/en/latest/review_checklist.html), a review isn't complete until all items of the checklist have been ticked off.) Thanks!
