### 2023-05-02 #5312 @plaplant
@mwalmsley looking at the references in the paper, it seems like [the reference for Phan et al.](https://github.com/mwalmsley/zoobot/blob/d9e8c2c3152c78acabfc8cf4c567aea97ce4164c/paper/paper.bib#L194) does not have a DOI associated with it. For this paper, you can find it [here](https://arxiv.org/abs/1912.11554), so please update the reference in `paper.bib` accordingly.

Also, it looks like [the recommended reference for Pyro](https://github.com/pyro-ppl/pyro#citation) is this:
```
@article{bingham2019pyro,
  author    = {Eli Bingham and
               Jonathan P. Chen and
               Martin Jankowiak and
               Fritz Obermeyer and
               Neeraj Pradhan and
               Theofanis Karaletsos and
               Rohit Singh and
               Paul A. Szerlip and
               Paul Horsfall and
               Noah D. Goodman},
  title     = {Pyro: Deep Universal Probabilistic Programming},
  journal   = {J. Mach. Learn. Res.},
  volume    = {20},
  pages     = {28:1--28:6},
  year      = {2019},
  url       = {http://jmlr.org/papers/v20/18-403.html}
}
```
It does not seem like the journal version has a DOI associated with it, but I think this is still the better version to cite (as opposed to the arXiv version). Thanks!
