### 2023-05-04 #5312 @mwalmsley
Thanks for the references note @plaplant. I've added the DOI to Phan et al and updated the Pyro reference to the suggested reference.

I've created a new Zoobot tagged release, 1.0.3, which is versioned on Zenodo [here](https://zenodo.org/record/7896025#.ZFOZmnaZMTU).
