### 2023-05-04 #5312 @plaplant
@mwalmsley thanks for the info!

One last question: I notice that not all of the authors have an ORCID associated with them. Does everyone in the author list who has an ORCID have one listed in the paper metadata? (Note that not all authors need an ORCID, just wanted to make sure they were linked if they had one.) Also, please add author ORCIDs to the Zenodo metadata as well. Thanks very much!
