+++ paper/paper.bib
@@ -159,6 +159,7 @@ archivePrefix = {arXiv},
    title = {On the Opportunities and Risks of Foundation Models},
    url = {http://arxiv.org/abs/2108.07258},
    year = {2021},
+   doi = {10.48550/arXiv.2108.07258}
 }
 
 @misc{https://doi.org/10.48550/arxiv.2104.10972,
