### 2023-05-06 #5312 @plaplant
Thanks! I think there are just a few outstanding issues:
1. Please confirm that all of the co-authors with ORCIDs are the only ones who have one.
2. Please add the co-authors' ORCIDs to the Zenodo metadata.
3. Please adjust the title of the Zenodo archive to match the name of the paper ("Zoobot: Adaptable Deep Learning Models for Galaxy Morphology").

At that point I think the paper is ready for acceptance. Let me know if you have any questions!
