### 2023-05-07 #5312 @mwalmsley
Brilliant, thank you!

1. I've added three additional ORCIDs following a manual search. The remaining co-authors without ORCIDs did not tell me their ORCID (only affiliations) and I can't identify any ORCID from the database.
2/3 I've updated the Zenodo details to match the JOSS paper details (it was first created last year as a simple code archive)
