### 2023-05-07 #5312 @plaplant
@mwalmsley it looks like [the reference to Pyro in the `paper.md` file](https://github.com/mwalmsley/zoobot/blob/ca3fb1a862d7bf4d271590627592281afcf458f3/paper/paper.md?plain=1#L194) did not get updated to [the new citation key in `paper.bib`](https://github.com/mwalmsley/zoobot/blob/ca3fb1a862d7bf4d271590627592281afcf458f3/paper/paper.bib#L186). I think once that is updated, this error should go away.
