### 2023-05-07 #5312 @mwalmsley
Ah, didn't notice the year changed - should be fixed now.

On Sun, May 7, 2023 at 6:21 PM Paul La Plante ***@***.***>
wrote:

> @mwalmsley <https://github.com/mwalmsley> it looks like the reference to
> Pyro in the paper.md file
> <https://github.com/mwalmsley/zoobot/blob/ca3fb1a862d7bf4d271590627592281afcf458f3/paper/paper.md?plain=1#L194>
> did not get updated to the new citation key in paper.bib
> <https://github.com/mwalmsley/zoobot/blob/ca3fb1a862d7bf4d271590627592281afcf458f3/paper/paper.bib#L186>.
> I think once that is updated, this error should go away.
>
