### 2023-05-08 #5312 @dfm
I got the bot mixed up by asking it to accept twice, but those error messages are not correct... here's what it should have said (note that while the DOI already resolves it may take a little while for the actual manuscript to show up because of caching issues):

Here's what you must now do:
